+++ curvlinops/__init__.py
@@ -1,6 +1,8 @@
 """``curvlinops`` library API."""
 
+from curvlinops.fisher import FisherMCLinearOperator
 from curvlinops.ggn import GGNLinearOperator
+from curvlinops.gradient_moments import EFLinearOperator
 from curvlinops.hessian import HessianLinearOperator
 from curvlinops.inverse import CGInverseLinearOperator
 from curvlinops.papyan2020traces.spectrum import (
@@ -9,11 +11,15 @@ from curvlinops.papyan2020traces.spectrum import (
     lanczos_approximate_log_spectrum,
     lanczos_approximate_spectrum,
 )
+from curvlinops.submatrix import SubmatrixLinearOperator
 
 __all__ = [
     "HessianLinearOperator",
     "GGNLinearOperator",
+    "EFLinearOperator",
+    "FisherMCLinearOperator",
     "CGInverseLinearOperator",
+    "SubmatrixLinearOperator",
     "lanczos_approximate_spectrum",
     "lanczos_approximate_log_spectrum",
     "LanczosApproximateSpectrumCached",