+++ curvlinops/ggn.py
@@ -9,7 +9,34 @@ from curvlinops._base import _LinearOperator
 
 
 class GGNLinearOperator(_LinearOperator):
-    """GGN as SciPy linear operator."""
+    r"""GGN as SciPy linear operator.
+
+    Consider the empirical risk
+
+    .. math::
+        \mathcal{L}(\mathbf{\theta})
+        =
+        c \sum_{n=1}^{N}
+        \ell(f_{\mathbf{\theta}}(\mathbf{x}_n), \mathbf{y}_n)
+
+    with :math:`c = \frac{1}{N}` for ``reduction='mean'`` and :math:`c=1` for
+    ``reduction='sum'``. The GGN matrix is
+
+    .. math::
+        c \sum_{n=1}^{N}
+        \left(
+            \mathbf{J}_{\mathbf{\theta}}
+            f_{\mathbf{\theta}}(\mathbf{x}_n)
+        \right)^\top
+        \left(
+            \nabla_{f_\mathbf{\theta}(\mathbf{x}_n)}^2
+            \ell(f_{\mathbf{\theta}}(\mathbf{x}_n), \mathbf{y}_n)
+        \right)
+        \left(
+            \mathbf{J}_{\mathbf{\theta}}
+            f_{\mathbf{\theta}}(\mathbf{x}_n)
+        \right)\,.
+    """
 
     def _matvec_batch(
         self, X: Tensor, y: Tensor, x_list: List[Tensor]